chg: fix deletecar controller to check the car exist or not and fix Readme file

+++ cars-management/src/controller/deleteCars.ts
@@ -6,6 +6,14 @@ import { Car} from "../models/cars"
 const deleteCar = async (req: Request, res: Response) => {
     const { id } = req.params;
 
+    // At first get car data from Car collection in database by id
+    const car = await Car.findOne({ _id: id });
+
+    // if there is no car according to car id send status Not Found
+    if (!car) {
+        return res.status(404).json({ message: `Car with id "${id}" not found.` });
+    }
+
     // Delete the car data according the car id
     await Car.findByIdAndDelete(id);
